+++ lib/ecto_foundationdb/tenant.ex
@@ -33,6 +33,7 @@ defmodule EctoFoundationDB.Tenant do
 
   alias Ecto.Adapters.FoundationDB, as: FDB
 
+  alias EctoFoundationDB.Layer.PrimaryKVCodec
   alias EctoFoundationDB.Migrator
   alias EctoFoundationDB.Options
   alias EctoFoundationDB.Tenant
@@ -163,9 +164,13 @@ defmodule EctoFoundationDB.Tenant do
   keyspace.
   """
   def pack(tenant, tuple) when is_tuple(tuple) do
-    tuple
-    |> tenant.backend.extend_tuple(tenant.meta)
-    |> :erlfdb_tuple.pack()
+    tuple = tenant.backend.extend_tuple(tuple, tenant.meta)
+    :erlfdb_tuple.pack(tuple)
+  end
+
+  def primary_codec(tenant, tuple) when is_tuple(tuple) do
+    tuple = tenant.backend.extend_tuple(tuple, tenant.meta)
+    PrimaryKVCodec.new(tuple)
   end
 
   def unpack(tenant, tuple) do